Move shared views to new common dir
