+++ lib/assets/javascripts/cartodb/new_common/dialogs/delete_items_view.js
@@ -1,8 +1,8 @@
 var cdb = require('cartodb.js');
-var BaseDialog = require('../../new_common/views/base_dialog/view');
-var pluralizeString = require('../../new_common/view_helpers/pluralize_string');
-var randomQuote = require('../../new_common/view_helpers/random_quote');
-var MapCardPreview = require('../../new_dashboard/mapcard_preview');
+var BaseDialog = require('../views/base_dialog/view');
+var pluralizeString = require('../view_helpers/pluralize_string');
+var randomQuote = require('../view_helpers/random_quote');
+var MapCardPreview = require('../views/mapcard_preview');
 var _ = require('underscore');
 var $ = require('jquery');
 var moment = require('moment');
@@ -56,14 +56,14 @@ module.exports = BaseDialog.extend({
   },
 
   _renderLoadingPrerequisites: function() {
-    return cdb.templates.getTemplate('new_dashboard/templates/loading')({
+    return cdb.templates.getTemplate('new_common/templates/loading')({
       title: 'Checking what consequences deleting the selected ' + this._pluralizedContentType() + ' would have...',
       quote: randomQuote()
     });
   },
 
   _renderLoadPrerequisitesFail: function() {
-    return cdb.templates.getTemplate('new_dashboard/templates/fail')({
+    return cdb.templates.getTemplate('new_common/templates/fail')({
       msg: 'Failed to check consequences of deleting the selected ' + this._pluralizedContentType()
     });
   },
@@ -73,7 +73,7 @@ module.exports = BaseDialog.extend({
     var affectedEntities = this._viewModel.affectedEntities();
     var affectedVisData = this._viewModel.affectedVisData();
 
-    return cdb.templates.getTemplate('new_dashboard/dialogs/delete_items_view_template')({
+    return cdb.templates.getTemplate('new_common/dialogs/delete_items_view_template')({
       selectedCount: this._viewModel.length,
       isDatasets: this._viewModel.isDeletingDatasets(),
       pluralizedContentType: this._pluralizedContentType(),
@@ -130,14 +130,14 @@ module.exports = BaseDialog.extend({
   },
 
   _renderDeletingItems: function() {
-    return cdb.templates.getTemplate('new_dashboard/templates/loading')({
+    return cdb.templates.getTemplate('new_common/templates/loading')({
       title: 'Deleting the selected ' + this._pluralizedContentType() + '...',
       quote: randomQuote()
     });
   },
 
   _renderDeleteItemsFail: function() {
-    return cdb.templates.getTemplate('new_dashboard/templates/fail')({
+    return cdb.templates.getTemplate('new_common/templates/fail')({
       msg: 'Failed to delete the selected ' + this._pluralizedContentType()
     });
   },
