+++ lib/assets/javascripts/cartodb/new_common/dialogs/delete_items_view_model.js
