+++ lib/assets/javascripts/cartodb/new_common/dialogs/delete_items_view_template.jst.ejs
