+++ lib/assets/javascripts/cartodb/new_common/templates/fail.jst.ejs
