+++ lib/assets/javascripts/cartodb/new_common/templates/loading.jst.ejs
