+++ lib/assets/javascripts/cartodb/new_common/views/mapcard_preview.js
