+++ lib/assets/javascripts/cartodb/new_dashboard/datasets/loading_library_view.js
@@ -7,7 +7,7 @@ module.exports = cdb.core.View.extend({
 
   render: function() {
     this.$el.html(
-      cdb.templates.getTemplate('new_dashboard/templates/loading')({
+      cdb.templates.getTemplate('new_common/templates/loading')({
         title: 'There are no datasets in the library just yet',
         quote: 'Your library is being stocked up as we speak, please come back in a few moments!'
       })
