+++ lib/assets/javascripts/cartodb/new_dashboard/dialogs/change_lock_view.js
@@ -70,7 +70,7 @@ module.exports = BaseDialog.extend({
 
   _renderProcessingItems: function() {
     var lockingOrUnlockingStr = this._viewModel.initialLockValue() ? 'Unlocking' : 'Locking';
-    return cdb.templates.getTemplate('new_dashboard/templates/loading')({
+    return cdb.templates.getTemplate('new_common/templates/loading')({
       title: lockingOrUnlockingStr + ' ' + pluralizeString(this.options.contentType === 'datasets' ? 'dataset' : 'map', this._viewModel.length) + '...',
       quote: randomQuote()
     });
@@ -78,7 +78,7 @@ module.exports = BaseDialog.extend({
 
   _renderProcessItemsFail: function() {
     var lockingOrUnlockingStr = this._viewModel.initialLockValue() ? 'unlock' : 'lock';
-    return cdb.templates.getTemplate('new_dashboard/templates/fail')({
+    return cdb.templates.getTemplate('new_common/templates/fail')({
       msg: 'Failed to ' + lockingOrUnlockingStr + ' all items'
     });
   }
