+++ lib/assets/javascripts/cartodb/new_dashboard/dialogs/change_privacy_view.js
@@ -92,14 +92,14 @@ module.exports = BaseDialog.extend({
   },
 
   _renderSaving: function() {
-    return cdb.templates.getTemplate('new_dashboard/templates/loading')({
+    return cdb.templates.getTemplate('new_common/templates/loading')({
       title: 'Saving privacy...',
       quote: randomQuote()
     });
   },
 
   _renderSaveFail: function() {
-    return cdb.templates.getTemplate('new_dashboard/templates/fail')({
+    return cdb.templates.getTemplate('new_common/templates/fail')({
       msg: ''
     });
   },
