+++ lib/assets/javascripts/cartodb/new_dashboard/filters_view.js
@@ -4,10 +4,10 @@ var _ = require('underscore');
 var Utils = require('cdb.Utils');
 var navigateThroughRouter = require('../new_common/view_helpers/navigate_through_router');
 var pluralizeString = require('../new_common/view_helpers/pluralize_string');
-var DeleteItemsDialog = require('../new_dashboard/dialogs/delete_items_view');
+var DeleteItemsDialog = require('../new_common/dialogs/delete_items_view');
 var ChangeLockDialog = require('../new_dashboard/dialogs/change_lock_view');
 var ChangeLockViewModel = require('./dialogs/change_lock_view_model');
-var DeleteItemsViewModel = require('./dialogs/delete_items_view_model');
+var DeleteItemsViewModel = require('../new_common/dialogs/delete_items_view_model');
 
 /**
  *  Dashboard filters.
