+++ lib/assets/javascripts/cartodb/new_dashboard/maps/maps_item.js
@@ -2,7 +2,7 @@ var cdb = require('cartodb.js');
 var moment = require('moment');
 var navigateThroughRouter = require('../../new_common/view_helpers/navigate_through_router');
 var MapviewsGraph = require('../../new_dashboard/mapviews_graph');
-var MapCardPreview = require('../../new_dashboard/mapcard_preview');
+var MapCardPreview = require('../../new_common/views/mapcard_preview');
 var LikesView = require('../../new_common/views/likes/view');
 var Utils = require('cdb.Utils');
 cdb.admin = require('cdb.admin');
@@ -129,7 +129,7 @@ module.exports = cdb.core.View.extend({
       this.killEvent(e);
       var isOwner = this.model.permission.isOwner(this.user);
       if (isOwner) {
-        this.model.set('selected', !this.model.get('selected'));  
+        this.model.set('selected', !this.model.get('selected'));
       }
     }
   }
