+++ lib/assets/javascripts/cartodb/new_public_dashboard/entry.js
@@ -5,7 +5,7 @@ var UserInfoView = require('./user_info_view');
 var PaginationModel = require('../new_common/views/pagination/model');
 var PaginationView = require('../new_common/views/pagination/view');
 var UserSettingsView = require('./user_settings_view');
-var MapCardPreview = require('../new_dashboard/mapcard_preview');
+var MapCardPreview = require('../new_common/views/mapcard_preview');
 var LikeView = require('../new_common/views/likes/view');
 
 $(function() {
