+++ lib/assets/javascripts/cartodb/new_table.js
@@ -2,5 +2,5 @@
 // usable in a non-browserified bundle
 //
 // Expected to be loaded after cdb.js but before table.js
-cdb.admin.DeleteItemsView      = require('./new_dashboard/dialogs/delete_items_view');
-cdb.admin.DeleteItemsViewModel = require('./new_dashboard/dialogs/delete_items_view_model');
+cdb.admin.DeleteItemsView      = require('./new_common/dialogs/delete_items_view');
+cdb.admin.DeleteItemsViewModel = require('./new_common/dialogs/delete_items_view_model');
