+++ lib/assets/test/spec/cartodb/new_common/dialogs/delete_items_view.spec.js
@@ -1,9 +1,9 @@
 // Mock default behaviour for dependency, re-apply explicitly for tests where we want to test this mixin.
-var DeleteItems = require('../../../../../javascripts/cartodb/new_dashboard/dialogs/delete_items_view');
-var ViewModel = require('../../../../../javascripts/cartodb/new_dashboard/dialogs/delete_items_view_model');
+var DeleteItems = require('../../../../../javascripts/cartodb/new_common/dialogs/delete_items_view');
+var ViewModel = require('../../../../../javascripts/cartodb/new_common/dialogs/delete_items_view_model');
 var cdbAdmin = require('cdb.admin');
 
-describe('new_dashboard/dialogs/delete_items_view', function() {
+describe('new_common/dialogs/delete_items_view', function() {
   beforeEach(function() {
     this.user = new cdbAdmin.User({
       base_url: 'http://pepe.cartodb.com',
