+++ lib/assets/test/spec/cartodb/new_common/dialogs/delete_items_view_model.spec.js
@@ -1,8 +1,8 @@
-var DeleteItemsViewModel = require('../../../../../javascripts/cartodb/new_dashboard/dialogs/delete_items_view_model');
+var DeleteItemsViewModel = require('../../../../../javascripts/cartodb/new_common/dialogs/delete_items_view_model');
 var cdb = require('cartodb.js');
 var $ = require('jquery');
 
-describe('new_dashboard/dialogs/delete_items_view_model', function() {
+describe('new_common/dialogs/delete_items_view_model', function() {
   beforeEach(function() {
     this.models = [
       new cdb.admin.Visualization({}),
