+++ lib/assets/test/spec/cartodb/new_common/views/mapcard_preview.spec.js
@@ -1,11 +1,11 @@
 var Backbone = require('backbone');
 var cdb = require('cartodb.js');
 var $ = require('jquery');
-var MapCardPreview = require('../../../../javascripts/cartodb/new_dashboard/mapcard_preview');
+var MapCardPreview = require('../../../../../javascripts/cartodb/new_common/views/mapcard_preview');
 
 var view, custom_dimensions_view, fakeSpy;
 
-describe('new_dashboard/mapcard_preview', function() {
+describe('new_common/views/mapcard_preview', function() {
 
   beforeEach(function() {
