+++ lib/assets/test/spec/cartodb/new_dashboard/filters_view.spec.js
@@ -1,4 +1,4 @@
-var DeleteItemsDialog = require('../../../../javascripts/cartodb/new_dashboard/dialogs/delete_items_view');
+var DeleteItemsDialog = require('../../../../javascripts/cartodb/new_common/dialogs/delete_items_view');
 var ChangeLockDialog = require('../../../../javascripts/cartodb/new_dashboard/dialogs/change_lock_view');
 var FiltersView = require('../../../../javascripts/cartodb/new_dashboard/filters_view');
 var Router = require('../../../../javascripts/cartodb/new_dashboard/router');
